hotel_booking_system/book_hotel_service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydanticz
python-dotenv
SQLAlchemy-Utils
psycopg2-binary
pyjwt
dotenv
passlib[bcrypt]